airflow/requirements.txt
requests
pendulum
boto3
pyyaml
python-dotenv